# openapi/platform-api.json
{
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {},
          "fieldErrors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "type": "object"
      },
...
      "CreateKnowledgeRequest": {
        "properties": {
          "connectionId": {
            "format": "uuid",
            "type": "string"
          },
          "externalResourceId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "ownerDepartmentId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "connectionId",
          "name",
          "externalResourceId",
          "ownerDepartmentId"
        ],
        "type": "object"
      },
...
      "FieldError": {
        "properties": {
          "code": {
            "type": "string"
          },
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "UpdateExternalResourceRequest": {
        "properties": {
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "name",
          "status",
          "version"
        ],
        "type": "object"
      },
...
    },
...
  },
  "info": {
    "description": "",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "Agentx Platform API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/knowledge/resources": {
      "get": {
        "operationId": "list_knowledge",
        "responses": {},
        "tags": [
          "external_resources"
        ]
      },
      "post": {
        "operationId": "create_knowledge",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKnowledgeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {},
        "tags": [
          "external_resources"
        ]
      }
    },
    "/api/v1/knowledge/resources/{id}": {
      "delete": {
        "operationId": "delete_knowledge",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expectedVersion",
            "required": true,
            "schema": {
              "format": "int64",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "deletion"
        ]
      },
      "get": {
        "operationId": "get_knowledge",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": [
          "external_resources"
        ]
      },
      "patch": {
        "operationId": "update_knowledge",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExternalResourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {},
        "tags": [
          "external_resources"
        ]
      }
    },
...
  },
...
}